pyproject.toml
[tool.poetry]
name = "vidmergebot"
version = "1.5.3"
description = ""
authors = ["Divanshu Chauhan <23524935+Divkix@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asyncio = "^3.4.3"
Pyrogram = "^2.0.27"
TgCrypto = "^1.2.3"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
aiohttp = "^3.8.1"
hachoir = "^3.1.3"
humanfriendly = "^10.0"
Pillow = "^9.1.1"
cachetools = "^5.2.0"
pymongo = "^4.1.1"
dnspython = "^2.2.1"
aiofiles = "^0.8.0"
pyromod = {git = "https://github.com/usernein/pyromod"}
prettyconf = "^2.2.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.14.1"
pre-commit = "^2.19.0"
pyclean = "^2.2.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
